Fixed map legend
