--- components/explore-map/legend/index.js
@@ -1,1 +1,0 @@
-export { default } from './component';
