+++ package.json
@@ -17,6 +17,7 @@
     "@next/bundle-analyzer": "^9.3.6",
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
+    "ajv-keywords": "^3.5.1",
     "autoprefixer": "^9.7.6",
     "axios": "^0.19.2",
     "axios-cache-adapter": "^2.5.0",
